fix: 🐛 dom pollution
