+++ src/components/Select/view/style/Select.style.tsx
@@ -1,207 +1,226 @@
 import styled, { css } from 'styled-components';
 import { StyledSelectProps } from '../../types';
 import { theme } from '../../../../theme';
+import { excludeIntrinsicElementProps } from '../../../../utils/excludeProps';
 
-export const SelectContainer = styled.div<StyledSelectProps>`
-  position: relative;
-  width: 220px;
-  line-height: 1;
+const selectIntrinsicExclude = excludeIntrinsicElementProps<
+    StyledSelectProps
+>();
+const selectPropKeys: (keyof StyledSelectProps)[] = [
+    'size',
+    'isDisabled',
+    'isFocused',
+    'isFullWidth',
+    'isMenuOpened',
+    'isMulti',
+    'size',
+];
 
-  font-size: ${(props) => props.theme.font.size.md};
-  font-family: ${(props) => props.theme.font.family};
+export const SelectContainer = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
+    position: relative;
+    width: 220px;
+    line-height: 1;
 
-  * {
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-  }
- 
-  ${(props) =>
-      props.isFullWidth &&
-      css`
-          width: 100%;
-      `}
-  
-  ${(props) =>
-      props.size === 'small' &&
-      css`
-          font-size: ${(props) => props.theme.font.size.sm};
-      `}
-  
-  ${(props) =>
-      props.size === 'large' &&
-      css`
-          font-size: ${(props) => props.theme.font.size.lg};
-      `}
-`;
+    font-size: ${(props) => props.theme.font.size.md};
+    font-family: ${(props) => props.theme.font.family};
 
-export const Control = styled.div<StyledSelectProps>`
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  padding-left: ${(props) => props.theme.padding.md};
-  border: 1px solid ${(props) => props.theme.color.border};
-  border-radius: ${(props) => props.theme.radius};
-  background-color: ${(props) => props.theme.color.background.alpha};
-  
-  :hover{
-    border: 1px solid ${(props) => props.theme.color.primary.alpha};
-  }
-  
-  ${(props) =>
-      props.isFocused &&
-      css`
-          border-color: ${(props) => props.theme.color.primary.alpha};
-      `}
-  
-  ${(props) =>
-      !props.isFocused &&
-      css`
-          border-color: ${(props) => props.theme.color.border};
-      `}
-  
-  ${(props) =>
-      props.isDisabled &&
-      css`
-          color: ${props.theme.color.text.beta};
-
-          ::placeholder {
-              color: ${props.theme.color.text.beta};
-          }
-
-          background-color: ${props.theme.color.default};
-          cursor: not-allowed;
-          outline: 0;
-
-          &:hover,
-          &:focus {
-              outline: 0;
-              box-shadow: none;
-              border: 1px solid ${props.theme.color.border};
-          }
-      `}
-  
-  ${(props) =>
-      props.isMenuOpened &&
-      css`
-          border-bottom: 1px solid ${(props) => props.theme.color.border};
-          border-bottom-left-radius: 0;
-          border-bottom-right-radius: 0;
-      `}
-  
-  ${(props) =>
-      props.isMulti &&
-      props.hasValue &&
-      css`
-          padding-left: 0;
-      `}
-  
-  ${(props) =>
-      props.isFullWidth &&
-      css`
-          width: 100%;
-      `}
-`;
-
-export const SelectContainerWrapper = styled.span<any>`
-   font-family: ${(props) => props.theme.font.family};
-  display: inline-block;
-  position: relative;
-  margin-top: -${(props) => (props.isMulti ? '18px' : props.theme.margin.md)};
-  padding-top: ${(props) => (props.isMulti ? '18px' : props.theme.padding.md)};
-  
-   ${(props) =>
-       props.isFocused &&
-       css`
-           ${Control} {
-               border: 1px solid ${(props) => props.theme.color.primary.alpha};
-           }
-       `}
-   
-   ${(props) =>
-       !props.isFocused &&
-       css`
-           ${Control} {
-               border: 1px solid ${(props) => props.theme.color.border};
-           }
-       `}
-  
-  :hover{
-    ${Control}{
-      border: 1px solid ${(props) => props.theme.color.primary.alpha};
+    * {
+        -webkit-box-sizing: border-box;
+        -moz-box-sizing: border-box;
+        box-sizing: border-box;
     }
-  }
-  
-    label {   
-    z-index: 1;
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    padding: 0 ${(props) => props.theme.padding.md};
-    border-radius: ${(props) => props.theme.radius};
 
-    color: ${(props) => props.theme.color.text.beta};
-    cursor: pointer;
-    border: ${(props) => props.theme.radius};
+    ${(props) =>
+        props.isFullWidth &&
+        css`
+            width: 100%;
+        `}
 
-    -webkit-transition: color 10s ease-in-out, transform 10s ease-in-out;
-    transition: color 10s ease-in-out, transform 10s ease-in-out;
-    -webkit-transform-origin: 0 100%;
-    transform-origin: 0 100%;
-    -webkit-transform:  ${(props) =>
-        props.isMulti ? 'translateY(18px)' : 'translateY(12px)'};
-    transform: ${(props) =>
-        props.isMulti ? 'translateY(18px)' : 'translateY(12px)'};
-    
     ${(props) =>
         props.size === 'small' &&
         css`
-            font-size: ${props.theme.font.size.sm};
-            height: ${parseInt(props.theme.height.sm, 0) - 2}px;
-            line-height: ${parseInt(props.theme.height.sm, 0) - 2}px;
+            font-size: ${(props) => props.theme.font.size.sm};
         `}
-    ${(props) =>
-        props.size === 'medium' &&
-        css`
-            font-size: ${props.theme.font.size.md};
-            height: ${parseInt(props.theme.height.md, 0) - 2}px;
-            line-height: ${parseInt(props.theme.height.md, 0) - 2}px;
-        `}
-    ${(props) =>
+  
+  ${(props) =>
         props.size === 'large' &&
         css`
-            font-size: ${props.theme.font.size.lg};
-            height: ${parseInt(props.theme.height.lg, 0) - 2}px;
-            line-height: ${parseInt(props.theme.height.lg, 0) - 2}px;
+            font-size: ${(props) => props.theme.font.size.lg};
+        `}
+`;
+
+export const Control = styled(selectIntrinsicExclude('div', selectPropKeys))`
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    cursor: pointer;
+    padding-left: ${(props) => props.theme.padding.md};
+    border: 1px solid ${(props) => props.theme.color.border};
+    border-radius: ${(props) => props.theme.radius};
+    background-color: ${(props) => props.theme.color.background.alpha};
+
+    :hover {
+        border: 1px solid ${(props) => props.theme.color.primary.alpha};
+    }
+
+    ${(props) =>
+        props.isFocused &&
+        css`
+            border-color: ${(props) => props.theme.color.primary.alpha};
         `}
 
     ${(props) =>
-        (props.isFocused || props.hasValue) &&
+        !props.isFocused &&
         css`
-            height: ${props.theme.height.sm};
-            line-height: ${props.theme.height.sm};
-            left: 5px;
-            right: auto;
-
-            background: ${props.theme.color.background.alpha};
-            color: ${props.theme.color.primary.alpha};
-
-            -webkit-transform: translateY(-14px) scale(0.8);
-            transform: translateY(0px) scale(0.8);
-            -webkit-transform-origin: 0 0;
-            transform-origin: 0 0;
-
-            ${props.size === 'small' &&
-            css`
-                height: 16px;
-                line-height: 16px;
-            `}
+            border-color: ${(props) => props.theme.color.border};
         `}
-  }
-    
+  
+  ${(props) =>
+        props.isDisabled &&
+        css`
+            color: ${props.theme.color.text.beta};
+
+            ::placeholder {
+                color: ${props.theme.color.text.beta};
+            }
+
+            background-color: ${props.theme.color.default};
+            cursor: not-allowed;
+            outline: 0;
+
+            &:hover,
+            &:focus {
+                outline: 0;
+                box-shadow: none;
+                border: 1px solid ${props.theme.color.border};
+            }
+        `}
+  
+  ${(props) =>
+        props.isMenuOpened &&
+        css`
+            border-bottom: 1px solid ${(props) => props.theme.color.border};
+            border-bottom-left-radius: 0;
+            border-bottom-right-radius: 0;
+        `}
+  
+  ${(props) =>
+        props.isMulti &&
+        props.hasValue &&
+        css`
+            padding-left: 0;
+        `}
+  
+  ${(props) =>
+        props.isFullWidth &&
+        css`
+            width: 100%;
+        `}
+`;
+
+export const SelectContainerWrapper = styled(
+    selectIntrinsicExclude('span', selectPropKeys)
+)`
+    font-family: ${(props) => props.theme.font.family};
+    display: inline-block;
+    position: relative;
+    margin-top: -${(props) => (props.isMulti ? '18px' : props.theme.margin.md)};
+    padding-top: ${(props) =>
+        props.isMulti ? '18px' : props.theme.padding.md};
+
+    ${(props) =>
+        props.isFocused &&
+        css`
+            ${Control} {
+                border: 1px solid ${(props) => props.theme.color.primary.alpha};
+            }
+        `}
+
+    ${(props) =>
+        !props.isFocused &&
+        css`
+            ${Control} {
+                border: 1px solid ${(props) => props.theme.color.border};
+            }
+        `}
+  
+  :hover {
+        ${Control} {
+            border: 1px solid ${(props) => props.theme.color.primary.alpha};
+        }
+    }
+
+    label {
+        z-index: 1;
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        padding: 0 ${(props) => props.theme.padding.md};
+        border-radius: ${(props) => props.theme.radius};
+
+        color: ${(props) => props.theme.color.text.beta};
+        cursor: pointer;
+        border: ${(props) => props.theme.radius};
+
+        -webkit-transition: color 10s ease-in-out, transform 10s ease-in-out;
+        transition: color 10s ease-in-out, transform 10s ease-in-out;
+        -webkit-transform-origin: 0 100%;
+        transform-origin: 0 100%;
+        -webkit-transform: ${(props) =>
+            props.isMulti ? 'translateY(18px)' : 'translateY(12px)'};
+        transform: ${(props) =>
+            props.isMulti ? 'translateY(18px)' : 'translateY(12px)'};
+
+        ${(props) =>
+            props.size === 'small' &&
+            css`
+                font-size: ${props.theme.font.size.sm};
+                height: ${parseInt(props.theme.height.sm, 0) - 2}px;
+                line-height: ${parseInt(props.theme.height.sm, 0) - 2}px;
+            `}
+        ${(props) =>
+            props.size === 'medium' &&
+            css`
+                font-size: ${props.theme.font.size.md};
+                height: ${parseInt(props.theme.height.md, 0) - 2}px;
+                line-height: ${parseInt(props.theme.height.md, 0) - 2}px;
+            `}
+    ${(props) =>
+            props.size === 'large' &&
+            css`
+                font-size: ${props.theme.font.size.lg};
+                height: ${parseInt(props.theme.height.lg, 0) - 2}px;
+                line-height: ${parseInt(props.theme.height.lg, 0) - 2}px;
+            `}
+
+    ${(props) =>
+            (props.isFocused || props.hasValue) &&
+            css`
+                height: ${props.theme.height.sm};
+                line-height: ${props.theme.height.sm};
+                left: 5px;
+                right: auto;
+
+                background: ${props.theme.color.background.alpha};
+                color: ${props.theme.color.primary.alpha};
+
+                -webkit-transform: translateY(-14px) scale(0.8);
+                transform: translateY(0px) scale(0.8);
+                -webkit-transform-origin: 0 0;
+                transform-origin: 0 0;
+
+                ${props.size === 'small' &&
+                css`
+                    height: 16px;
+                    line-height: 16px;
+                `}
+            `}
+    }
+
     ${(props) =>
         props.isFullWidth &&
         css`
@@ -209,7 +228,9 @@ export const SelectContainerWrapper = styled.span<any>`
         `}
 `;
 
-export const ValueContainer = styled.div<StyledSelectProps>`
+export const ValueContainer = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: flex;
     flex-wrap: wrap;
     align-items: center;
@@ -229,52 +250,55 @@ export const ValueContainer = styled.div<StyledSelectProps>`
         `}
 `;
 
-export const SingleValue = styled.div<StyledSelectProps>`
-  display: block;
-  max-width: 165px;
-  height: 32px;
-  padding: ${(props) => props.theme.padding.sm} ${(props) =>
-    props.theme.padding.sm} ${(props) => props.theme.padding.sm} 0;
-  
-  border-radius: ${(props) => props.theme.radius};
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  
-  span{
-    color: ${(props) => props.theme.color.text.alpha};
-  }
-  
-  ${(props) =>
-      props.isFullWidth &&
-      css`
-          max-width: calc(100% - 2px);
-      `}
-  
-  ${(props) =>
-      props.size === 'small' &&
-      css`
-          height: 24px;
-          padding: 3px 3px 3px 0;
+export const SingleValue = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
+    display: block;
+    max-width: 165px;
+    height: 32px;
+    padding: ${(props) => props.theme.padding.sm}
+        ${(props) => props.theme.padding.sm}
+        ${(props) => props.theme.padding.sm} 0;
 
-          span {
-              font-size: ${(props) => props.theme.font.size.sm};
-          }
-      `}
+    border-radius: ${(props) => props.theme.radius};
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
+
+    span {
+        color: ${(props) => props.theme.color.text.alpha};
+    }
+
+    ${(props) =>
+        props.isFullWidth &&
+        css`
+            max-width: calc(100% - 2px);
+        `}
+
+    ${(props) =>
+        props.size === 'small' &&
+        css`
+            height: 24px;
+            padding: 3px 3px 3px 0;
+
+            span {
+                font-size: ${(props) => props.theme.font.size.sm};
+            }
+        `}
   
   ${(props) =>
-      props.size === 'large' &&
-      css`
-          height: 38px;
-          padding: 7px 7px 7px 0;
+        props.size === 'large' &&
+        css`
+            height: 38px;
+            padding: 7px 7px 7px 0;
 
-          span {
-              font-size: ${(props) => props.theme.font.size.lg};
-          }
-      `}
+            span {
+                font-size: ${(props) => props.theme.font.size.lg};
+            }
+        `}
 `;
 
-export const MultiValue = styled.div<StyledSelectProps>`
+export const MultiValue = styled(selectIntrinsicExclude('div', selectPropKeys))`
     display: flex;
     max-width: calc(100% - 6px);
     height: 28px;
@@ -326,7 +350,9 @@ export const MultiValue = styled.div<StyledSelectProps>`
         `}
 `;
 
-export const ClearIndicator = styled.div<StyledSelectProps>`
+export const ClearIndicator = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: flex;
     align-items: center;
     justify-content: center;
@@ -355,7 +381,9 @@ export const MultiValueLabel = styled.div`
     padding: 0 ${(props) => props.theme.padding.sm};
 `;
 
-export const MultiValueRemove = styled.div<StyledSelectProps>`
+export const MultiValueRemove = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: flex;
     justify-content: center;
     align-items: center;
@@ -396,15 +424,15 @@ export const NoOptionsMessage = styled.div`
     cursor: pointer;
 `;
 
-export const CrossIcon = styled.button`
-    //
-`;
+export const CrossIcon = styled.button``;
 
 export const Placeholder = styled.div`
     color: ${(props) => props.theme.color.text.beta};
 `;
 
-export const IndicatorsContainer = styled.div<StyledSelectProps>`
+export const IndicatorsContainer = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: flex;
     -webkit-box-align: center;
     align-items: center;
@@ -437,7 +465,9 @@ export const IndicatorsContainer = styled.div<StyledSelectProps>`
         `}
 `;
 
-export const DropdownIndicator = styled.div<StyledSelectProps>`
+export const DropdownIndicator = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: flex;
     align-items: center;
     justify-content: center;
@@ -488,7 +518,9 @@ export const MenuList = styled.div`
     outline-offset: -2px;
 `;
 
-export const StyledOption = styled.div<StyledSelectProps>`
+export const StyledOption = styled(
+    selectIntrinsicExclude('div', selectPropKeys)
+)`
     display: block;
     padding: 0 ${(props) => props.theme.padding.md};
     height: ${(props) => props.theme.height.md};
